pyproject.toml
[project]
name = "quantum-ai-orchestrator"
version = "0.0.1"
description = "An AI-orchestrated control plane for hybrid quantum-classical workloads"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11,<3.13"
authors = [
    { name = "Ken Rollins" },
]

# Phase 0/1 runtime dependencies. The CUDA-Q stack runs in a container
# (nvcr.io/nvidia/cuda-quantum:0.9.x); the host venv carries everything
# that doesn't need the CUDA-Q runtime.
dependencies = [
    # Quantum + QEC stack
    "stim>=1.13,<2",
    "pymatching>=2.2,<3",
    "qutip>=5.0,<6",
    "dwave-neal>=0.6,<0.7",
    "dwave-inspector>=0.5,<0.6",

    # Optimization (classical)
    "ortools>=9.11,<10",

    # Numerics + data
    "numpy>=1.26,<2",
    "pandas>=2.2,<3",
    "scipy>=1.13,<2",

    # LLM + retrieval
    "openai>=1.40,<2",        # Ollama OpenAI-compatible client
    "litellm>=1.40,<2",       # optional unified gateway (host has it on :4000)

    # Storage + observability
    "psycopg[binary,pool]>=3.2,<4",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27",

    # Web replay server
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "sse-starlette>=2.1,<3",

    # Config + utilities
    "pydantic>=2.8,<3",
    "pyyaml>=6.0,<7",
    "typer>=0.12,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5.0,<6",
    "hypothesis>=6.111,<7",
    "ruff>=0.6,<1",
    "mypy>=1.11,<2",
]

# Optional GPU paths. Install only inside containers or where matching
# CUDA versions are present.
qiskit-gpu = [
    "qiskit>=1.2,<2",
    "qiskit-aer-gpu>=0.15,<1",
]

# Phase 2/3 deferred — install when those phases land.
phase2 = [
    "cuquantum-python>=24.08",  # cuTensorNet
]
phase3 = [
    "pennylane>=0.39,<1",
    "pennylane-lightning-gpu>=0.39,<1",
    "yfinance>=0.2,<1",
]

# Documentation site
docs = [
    "mkdocs>=1.6,<2",
    "mkdocs-material>=9.5,<10",
    "pymdown-extensions>=10.9,<11",
]

[project.scripts]
qao = "orchestrator.cli:main"
qao-replay = "orchestrator.cli:replay"
qao-bench = "orchestrator.cli:bench"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["orchestrator*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
